refactor: dgraph and graphs refactor (#1163)
* refactor: graph entities and remove dgraph interfaces

refactor: graph entities and remove dgraph interfaces

refactor: tag module tests

refactor: fields and finish type module

refactor: type module

refactor: element and type module refactor wip

refactor: refactor app module


refactor: remove dgraph interfaces


fix: jest setup


chore: merge with master

wip: element refactor

wip: element refactor

wip: codegen split into frontend & backend

wip: codegen split into frontend & backend

wip: refactor element data structure

wip: refactor element data structure

wip: refactor element data structure

wip: refactor element data structure

wip: refactor element data structure

wip: refactor element data structure

wip: refactor element data structure

wip: refactor element data structure

wip: refactor element data structure

wip: refactor element data structure

wip: refactor element data structure

wip: remove component use cases

wip: remove component use cases

wip: standardize interface schema

wip: standardize interface schema

wip: standardize interface schema

wip: standardize interface schema

wip: standardize interface schema

wip: standardize interface schema

wip: standardize interface schema

wip: standardize interface schema

fix: merge issues

chore: codegen

refactor: element tests

fix: page tests

refactor: element graph

fix: tests

fix: build

fix: build

fix: build

fix: build

* fix: tsconfig

* fix: remove codegen libs

* fix: lint

* fix: minor fixes

* fix: ci

* fix: ci

* fix: ci tests

* fix: ci

* revert: ci remove cache

* fix: ci

* revert: package updates

* revert: package updates

* fix: yarn lock

* fix: ci

* fix: yarn lock

* fix: e2e

* fix: e2e tests

* fix: enum type value order

* fix: seeder

* fix: dgraph rate limit issue

* fix: dgraph rate limit issue

* fix: dgraph rate limit issue

* fix: dgraph rate limit issue

* fix: dgraph rate limit issue

* fix: dgraph rate limit issue

* fix: ci quota issue

* fix: dgraph rate limit issue

* fix: user integration test

* fix: remove cypress retries

* fix: remove cypress retries

* fix: bad workspace config

* fix: cypress issue

* fix: switch cypress to firefox

* fix: switch to firefox in cypress ci

* fix: switch to firefox in cypress ci

* fix: remove type test from builder

* fix: revert to cypress 7

* revert: cypress retries

* test: remove cypress recording

* fix: cypress

* test: restructure builder props spec

* test: firefox in cypress ci

* fix: cypress dep bump

* fix: builder props test

* fix: builder props spec

* fix: e2e

* test: ignore troubling specs

* test: fix render props spec

+++ libs/backend/modules/lambda/src/application/lambda.resolver.ts
@@ -1,8 +1,6 @@
-import { Void } from '@codelab/backend/abstract/types'
-import { CreateResponse } from '@codelab/backend/application'
 import { GqlAuthGuard } from '@codelab/backend/infra'
 import { CurrentUser } from '@codelab/backend/modules/user'
-import type { User } from '@codelab/shared/abstract/core'
+import type { IUser } from '@codelab/shared/abstract/core'
 import { Injectable, UseGuards } from '@nestjs/common'
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql'
 import { Lambda, LambdaPayload } from '../domain/lambda.model'
@@ -21,7 +19,6 @@ import {
   UpdateLambdaInput,
   UpdateLambdaService,
 } from '../use-cases/update-lambda'
-import { LambdaAdapter } from './lambda.adapter'
 import { LambdaService } from './lambda.service'
 
 @Resolver(() => Lambda)
@@ -34,14 +31,13 @@ export class LambdaResolver {
     private readonly getLambdasService: GetLambdasService,
     private readonly getLambdaService: GetLambdaService,
     private readonly updateLambdaService: UpdateLambdaService,
-    private readonly lambdaAdapter: LambdaAdapter,
   ) {}
 
-  @Mutation(() => CreateResponse)
+  @Mutation(() => Lambda)
   @UseGuards(GqlAuthGuard)
   async createLambda(
     @Args('input') input: CreateLambdaInput,
-    @CurrentUser() currentUser: User,
+    @CurrentUser() currentUser: IUser,
   ) {
     const { id } = await this.createLambdaService.execute({
       input,
@@ -53,47 +49,73 @@ export class LambdaResolver {
       ...input,
     })
 
-    return { id }
+    const lambda = await this.getLambdaService.execute({
+      lambdaId: id,
+    })
+
+    if (!lambda) {
+      throw new Error("Can't find created lambda")
+    }
+
+    return lambda
   }
 
-  @Mutation(() => Void, { nullable: true })
+  @Mutation(() => Lambda)
   @UseGuards(GqlAuthGuard)
   async deleteLambda(@Args('input') input: DeleteLambdaInput) {
+    const lambda = await this.getLambdaService.execute({
+      lambdaId: input.lambdaId,
+    })
+
+    if (!lambda) {
+      throw new Error("Can't find lambda")
+    }
+
     await this.deleteLambdaService.execute(input)
 
-    return await this.lambdaService.deleteLambda({ id: input.lambdaId })
+    await this.lambdaService.deleteLambda({ id: input.lambdaId })
+
+    return lambda
   }
 
-  @Mutation(() => Void, { nullable: true })
+  @Mutation(() => Lambda, { nullable: true })
   @UseGuards(GqlAuthGuard)
   async updateLambda(@Args('input') input: UpdateLambdaInput) {
     await this.updateLambdaService.execute(input)
 
     await this.lambdaService.updateLambda(input)
+
+    const lambda = await this.getLambdaService.execute({
+      lambdaId: input.id,
+    })
+
+    if (!lambda) {
+      throw new Error("Can't find lambda")
+    }
+
+    return lambda
   }
 
   @Query(() => Lambda, { nullable: true })
   @UseGuards(GqlAuthGuard)
   async getLambda(@Args('input') input: GetLambdaInput) {
-    const dgraphLambda = await this.getLambdaService.execute(input)
+    const lambda = await this.getLambdaService.execute(input)
 
-    if (!dgraphLambda) {
+    if (!lambda) {
       return null
     }
 
-    return this.lambdaAdapter.mapItem(dgraphLambda)
+    return lambda
   }
 
   @Mutation(() => LambdaPayload, { nullable: true })
   @UseGuards(GqlAuthGuard)
   async executeLambda(@Args('input') input: ExecuteLambdaInput) {
-    const dgraphLambda = await this.getLambdaService.execute(input)
+    const lambda = await this.getLambdaService.execute(input)
 
-    if (!dgraphLambda) {
+    if (!lambda) {
       return null
     }
-
-    const lambda = this.lambdaAdapter.mapItem(dgraphLambda)
 
     const results = await this.lambdaService.executeLambda(
       lambda,
@@ -105,9 +127,7 @@ export class LambdaResolver {
 
   @Query(() => [Lambda], { defaultValue: [] })
   @UseGuards(GqlAuthGuard)
-  async getLambdas(@CurrentUser() currentUser: User) {
-    const dgraphLambdas = await this.getLambdasService.execute({ currentUser })
-
-    return this.lambdaAdapter.map(dgraphLambdas)
+  async getLambdas(@CurrentUser() currentUser: IUser) {
+    return this.getLambdasService.execute({ currentUser })
   }
 }
